refactor(stream): inline graceful shutdown cancellation
## What

- Inline graceful-drain cancellation alongside the HTTP handler paths and remove the now-redundant shared helper.
- Preserve cancellation, response-finalization, and request-body cleanup behavior for bidirectional requests.

## Why

- Keeping the lifecycle work where each handler owns its resources makes the distinct request-body cleanup path easier to maintain without changing the documented drain contract.

## Testing

- `make specs package=net/http/content/stream` - passed
- `make lint` - passed
- `make sec` - passed; govulncheck found no reachable vulnerabilities and Trivy found no repository vulnerabilities or secrets.
- Existing public handler tests cover draining before handler entry, during committed responses, and while a receive is blocked; CI will run the broader suite, fuzzing, benchmarks, and coverage.

AI-assisted-by: [Codex](https://openai.com/codex/)
